logging enhancements
